HIA-1277: Support decimals in scores endpoint

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsintegrationapi/models/assessRisksAndNeeds/ArnGroupReconvictionScore.kt
@@ -1,10 +1,11 @@
 package uk.gov.justice.digital.hmpps.hmppsintegrationapi.models.assessRisksAndNeeds
 
 import uk.gov.justice.digital.hmpps.hmppsintegrationapi.models.hmpps.GroupReconviction
+import java.math.BigDecimal
 
 data class ArnGroupReconvictionScore(
   val scoreLevel: String? = null,
-  val twoYears: Int? = null,
+  val twoYears: BigDecimal? = null,
 ) {
   fun toGroupReconviction(useV2NumericalValue: Boolean = true): GroupReconviction =
     GroupReconviction(
